refactor: crop studies and indicators weights

+++ src/modules/census/studies/services/crop-studies.ts
@@ -1,19 +1,15 @@
-import { Either, right } from "../../../../shared/Either";
+import { Either, left, right } from "../../../../shared/Either";
 import { CropStudies } from "../core/model/crop-studies";
-import { makeCensusStudiesRepository } from "../repositories/crop-studies.repository";
 import { ICensusStudiesRepository } from "../repositories/protocol/repository";
 
-export interface ICropStudiesServices {
+export interface ICreateCropStudiesService {
   create(params: {
     id_basin: number;
     data: Array<Omit<CropStudies, "id_basin">>;
   }): Promise<Either<Error, string>>
-  getByBasin(
-    id: number
-  ): Promise<Either<Error, { [k: string]: Omit<CropStudies, "crop"> } | null>>
 }
 
-class CropStudiesServices implements ICropStudiesServices {
+export class CreateCropStudiesService implements ICreateCropStudiesService {
   private readonly repository: ICensusStudiesRepository
 
   constructor(repository: ICensusStudiesRepository) {
@@ -24,20 +20,41 @@ class CropStudiesServices implements ICropStudiesServices {
     id_basin: number;
     data: Array<Omit<CropStudies, "id_basin">>;
   }): Promise<Either<Error, string>> {
-    // TO-DO: check use of batch update
-    // CHECK: is possible to create constraints in columns?
-    await this.repository.delete(
-      params.id_basin
-    );
+    const basinExists = await this.repository.checkIfBasinExists(params.id_basin)
 
-    console.log(`Sucesso ao deletar estudos da bacia ${params.id_basin}.`)
+    if (basinExists) {
+      // TO-DO: check use of batch update
+      // CHECK: is possible to create constraints in columns?
+      await this.repository.delete(
+        params.id_basin
+      );
 
-    await this.repository.create({
-      data: params.data,
-      id: params.id_basin
-    });
+      console.log(`Sucesso ao deletar estudos da bacia ${params.id_basin}.`)
 
-    return right(`Sucesso ao inserir estudos da bacia.`);
+      await this.repository.create({
+        data: params.data,
+        id: params.id_basin
+      });
+
+      return right(`Sucesso ao inserir estudos da bacia.`);
+    }
+
+    return left(new Error("Bacia não encontrada"))
+  }
+
+}
+
+export interface IGetCropStudiesByBasinService {
+  getByBasin(
+    id: number
+  ): Promise<Either<Error, { [k: string]: Omit<CropStudies, "crop"> } | null>>
+}
+
+export class GetCropStudiesByBasinService implements IGetCropStudiesByBasinService {
+  private readonly repository: ICensusStudiesRepository
+
+  constructor(repository: ICensusStudiesRepository) {
+    this.repository = repository
   }
 
   async getByBasin(
@@ -50,6 +67,3 @@ class CropStudiesServices implements ICropStudiesServices {
   }
 }
 
-
-
-export const makeCropStudiesServices = () => new CropStudiesServices(makeCensusStudiesRepository())
